requirements.txt
Flask==2.2.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4

